refactor: split monolith app into pages and components
