+++ src/app/apply/page.tsx
@@ -1,0 +1,8 @@
+import { ApplyPage, ApplicationInput } from '../../components/ApplyPage';
+
+export default function Apply() {
+  const handleSubmit = (a: ApplicationInput) => {
+    console.log(a);
+  };
+  return <ApplyPage onSubmit={handleSubmit} />;
+}
