+++ src/app/catalog/page.tsx
@@ -1,0 +1,7 @@
+import { Catalog, Product } from '../../components/Catalog';
+
+const sampleProducts: Product[] = [];
+
+export default function CatalogPage() {
+  return <Catalog products={sampleProducts} />;
+}
