+++ src/app/login/page.tsx
@@ -1,0 +1,6 @@
+import { LoginPage } from '../../components/LoginPage';
+
+export default function Login() {
+  const handleLogin = () => null;
+  return <LoginPage onLogin={handleLogin} onBack={() => {}} />;
+}
