+++ src/main.tsx
@@ -1,10 +1,2 @@
-// src/main.tsx
-import React from "react";
-import { createRoot } from "react-dom/client";
-import App from "./App";
-
-createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
+// Legacy entry point removed after migration to Next.js app directory.
+export {};
